+++ apps/lab/app/components/grainy-liquid/index.global.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import WebsiteLayout from './WebsiteLayout.vue'
+</script>
+
+<template>
+  <WebsiteLayout />
+</template>