+++ packages/nuxt/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.1.0 (2025-10-10)
+
+### 🧱 Updated Dependencies
+
+- Updated @tresjs/core to 5.1.0
+- Updated @tresjs/cientos to 5.1.0
+- Updated @tresjs/leches to 1.1.0
+
 ## 5.0.3 (2025-10-08)
 
 ### 🧱 Updated Dependencies